Add NoREC test for Postgres
